评论功能

+++ app/src/main/java/cn/mysilicon/merchant/OrderDetailActivity.java
@@ -11,11 +11,14 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
 import com.alibaba.fastjson.JSONArray;
 import com.bumptech.glide.Glide;
+
+import java.io.IOException;
 
 import cn.mysilicon.merchant.entity.Order;
 import okhttp3.Call;
@@ -25,6 +28,7 @@ import okhttp3.RequestBody;
 import okhttp3.Response;
 
 public class OrderDetailActivity extends AppCompatActivity {
+    private static final String TAG = "OrderDetailActivity";
     private Integer id;
     private Order order;
     private ImageView orderImage;
@@ -37,6 +41,7 @@ public class OrderDetailActivity extends AppCompatActivity {
     private TextView serverTime;
     private TextView orderStatus;
     private Button btnOrderDelete;
+    private Button btnOrderCancel;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -70,13 +75,49 @@ public class OrderDetailActivity extends AppCompatActivity {
         serverTime = findViewById(R.id.tv_order_server_time);
         orderStatus = findViewById(R.id.tv_order_status);
         btnOrderDelete = findViewById(R.id.btn_order_delete);
+        btnOrderCancel = findViewById(R.id.btn_order_cancel);
         if (order.getCur_status().equals("已完成")) {
             btnOrderDelete.setVisibility(Button.VISIBLE);
+            btnOrderCancel.setVisibility(Button.GONE);
             btnOrderDelete.setOnClickListener(v -> {
-                deleteOrder(order.getId());
+                AlertDialog.Builder builder = new AlertDialog.Builder(OrderDetailActivity.this);
+                builder.setTitle("提示");
+                builder.setMessage("确认删除该订单吗？");
+                builder.setPositiveButton("确认", (dialog, which) -> {
+                    deleteOrder(order.getId());
+                });
+                builder.setNegativeButton("取消", (dialog, which) -> {
+                });
+                builder.show();
             });
-        } else {
+        } if (order.getCur_status().equals("进行中")) {
             btnOrderDelete.setVisibility(Button.GONE);
+            btnOrderCancel.setVisibility(Button.VISIBLE);
+            btnOrderCancel.setOnClickListener(v -> {
+                AlertDialog.Builder builder = new AlertDialog.Builder(OrderDetailActivity.this);
+                builder.setTitle("提示");
+                builder.setMessage("确认取消该订单吗？");
+                builder.setPositiveButton("确认", (dialog, which) -> {
+                    cancleOrder(order.getId());
+                });
+                builder.setNegativeButton("取消", (dialog, which) -> {
+                });
+                builder.show();
+            });
+        }if (order.getCur_status().equals("已取消")) {
+            btnOrderDelete.setVisibility(Button.VISIBLE);
+            btnOrderCancel.setVisibility(Button.GONE);
+            btnOrderDelete.setOnClickListener(v -> {
+                AlertDialog.Builder builder = new AlertDialog.Builder(OrderDetailActivity.this);
+                builder.setTitle("提示");
+                builder.setMessage("确认删除该订单吗？");
+                builder.setPositiveButton("确认", (dialog, which) -> {
+                    deleteOrder(order.getId());
+                });
+                builder.setNegativeButton("取消", (dialog, which) -> {
+                });
+                builder.show();
+            });
         }
 
         Glide.with(this).load(order.getImage()).into(orderImage);
@@ -88,6 +129,34 @@ public class OrderDetailActivity extends AppCompatActivity {
         userAddress.setText(order.getAddress());
         serverTime.setText(order.getServer_time());
         orderStatus.setText(order.getCur_status());
+    }
+
+    private void cancleOrder(int id) {
+        new Thread(new Runnable() {
+            @Override
+            public void run() {
+                OkHttpClient client = new OkHttpClient();
+                Request request = new Request.Builder()
+                        .url("http://mysilicon.cn/order/cancel?id=" + id)
+                        .post(RequestBody.create("", null))
+                        .build();
+                Response response = null;
+                try {
+                    response = client.newCall(request).execute();
+                    Log.d(TAG, "run: " + response.body().string());
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+                if (response.code() != 200) {
+                    Looper.prepare();
+                    Toast.makeText(OrderDetailActivity.this, "网络错误", Toast.LENGTH_SHORT).show();
+                    Looper.loop();
+                } else {
+                    handler.sendEmptyMessage(2);
+                }
+
+            }
+        }).start();
     }
 
     private void deleteOrder(int id) {
@@ -129,6 +198,9 @@ public class OrderDetailActivity extends AppCompatActivity {
                 Toast.makeText(OrderDetailActivity.this, "删除成功", Toast.LENGTH_SHORT).show();
                 finish();
                 break;
+                case 2:
+                Toast.makeText(OrderDetailActivity.this, "取消成功", Toast.LENGTH_SHORT).show();
+                finish();
             default:
                 break;
         }
